registring api call gjort
