+++ index.html
@@ -7,7 +7,7 @@
     <meta name="keywords" content="" />
     <link rel="stylesheet" href="/css/styles.css" />
     <link rel="icon" href="data:;base64,=" />
-    <script src="./js/register.mjs" type="module"></script>
+    <script src="./js/scripts.mjs" type="module"></script>
     <title>The community</title>
   </head>
   <body>
