+++ js/scripts.mjs
@@ -1,3 +1,6 @@
-import { userName } from "./loginForm.mjs";
+import { postApiCall } from "./components/apiMethod.mjs";
+import { postData } from "./register.mjs";
+const API_BASE_URL = "https://nf-api.onrender.com";
+const registerUserUrl = `${API_BASE_URL}/api/v1/social/auth/register`;
 
-console.log(userName);
+// postApiCall(registerUserUrl, postData);
